Fixed one of the Serial.print commands in loop()
In the loop section of this sketch, there are a series of commands that print the RTC's time and date to the serial pins.  It is updated with the current time every second.  The text output is supposed to be (for example):
    Current RTC Date/Time: Sunday, April 23, 2017 10:10:10 (1492888806) 
When I ran the code after setting the clock, the output was more like:
   Current RTC Date/Time: December, April 22112962017 10:10:10 (1492888806)
The source of the non-sense number between month and year was traced one of the Serial.print commands that used single quotes for the string instead of double quotes.  I fixed it and the date appears to be printing correctly now.
This is a simple change that doesn't affect how the RTC is set or read.  It only changes the printed output for the code.