fix getTemperature for negative temperatures
Previously, getTemperature incorrectly treated the low byte of the 2s
complement temperature register, resulting in odd artifacts and incorrect
slopes in the fractional portion when the temperature is negative.

Refactor to use C++'s built in support for signed and unsigned integers
to handle this correctly and cleanly.

Tested on real hardware at negative temperatures.